Change TF version to cv32e40p port
Change the model of Tensorflow Lite for Microcontrol`lers used as the
ML framework for the benchmark to a fork that supports the CV32E40P.

As of this commit this port is hosted privately on a University of
Southampton in-house hosting of GitLab. This will hopefully change after
this project has been marked.