Add support for ML Accelerated CV32E40P
Add a new subdir to the bsp dir which adds support for the
ML Accelerated CV32E40P.

Remove the riscv-linux subdir from the bsp directory. It is replaced
with the addition of the makefile flag TAGS. This flag can be used to
specify build options. It is evaluated in the makefile.inc files in the
bsp subdirs.

The flag TAGS=spike" can be added to a call to make to
instruct the build system to ignore all linking instructions that enable
bare-metal execution on the CV32E40P. Without this information the
toolchain defaults to building TinyMLPerf to run on top of Linux.
This configuration is required for running software on Spike.